+++ src/test/java/domain/game/BlackjackGameManagerTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import dto.BlackjackResultDto;
 import factory.BlackjackControllerFactory;
-import constant.HitOrStand;
 import domain.card.Rank;
 import constant.Result;
 import domain.card.Suit;
